[#176236216] - Split add_stake off into submodule
